NP-2542 Move filters to common component
